Move marketplace data structure to seperate file
