+++ client/my-sites/marketplace/marketplace-product-definitions/test/test.marketplace-product-definitions.ts
@@ -2,6 +2,7 @@
  * External dependencies
  */
 import { YOAST_PREMIUM, YOAST_FREE } from '@automattic/calypso-products';
+
 /**
  * Internal dependencies
  */
@@ -11,8 +12,8 @@ import {
 	getDefaultProductInProductGroup,
 	getProductDefinition,
 	productGroups,
-} from '../constants';
-import { YOAST } from '../types';
+} from '../index';
+import { YOAST } from 'calypso/my-sites/marketplace/marketplace-product-definitions';
 
 describe( 'Constants and product related utilities test', () => {
 	test( 'getPluginsToInstall returns the relevant plugins to be installed', () => {
